project-container/requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.1.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
django-filter==2.2.0
pyyaml==5.3.1
uritemplate==3.0.1

